Fix
fix: Fixed typo referring to "Globe of Discordant Energy" as "Glove"
fix: Added checks in Shaman 1.0 quest to insure that Woe, Envy, and Marr's Promise are looted.
fix: Adjusted step numbers in Shaman 1.0 outline.
fix: Some renumbering of steps in Shaman 1.0 to accomodate the new verification steps.  If continuing Shaman 1.0 please select the step you are on in the Outline tab.
test: Additional work on Rogue 1.5 and Paladin 1.5. (Not yet fully implemented)